@@ -1,70 +1,55 @@
 {
     "name": "ddeg",
-    "version": "0.1.0",
+    "version": "1.0.0",
     "description": "DDEG - Dungeons & Dragons Encounter Generator for D&D 5e",
-    "homepage": "https://hayleyashby7.github.io/DDEG",
     "private": true,
+    "type": "module",
+    "scripts": {
+        "predeploy": "npm run build",
+        "deploy": "gh-pages -d dist",
+        "dev": "vite --open",
+        "build": "tsc && vite build",
+        "test": "jest ./src --verbose --coverage",
+        "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
+    },
     "dependencies": {
         "@fortawesome/fontawesome-svg-core": "^6.4.0",
         "@fortawesome/free-solid-svg-icons": "^6.4.0",
         "@fortawesome/react-fontawesome": "^0.2.0",
         "@hookform/error-message": "^2.0.1",
-        "@testing-library/jest-dom": "^5.16.5",
-        "@testing-library/react": "^14.0.0",
-        "@testing-library/user-event": "^14.0.0",
+        "@types/react-router-dom": "^5.3.3",
         "@vitejs/plugin-react": "^3.1.0",
-        "prop-types": "^15.8.1",
+        "axios": "^1.4.0",        
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-hook-form": "^7.43.9",
-        "vite": "^4.2.1",
-        "vite-plugin-eslint": "^1.8.1",
-        "vitest": "^0.30.0"
-    },
-    "scripts": {
-        "predeploy": "npm run build",
-        "deploy": "gh-pages -d dist",
-        "dev": "vite",
-        "build": "vite build && tailwind -i src/index.css -o dist/output.css",
-        "test": "vitest",
-        "coverage": "vitest run --coverage",
-        "lint": "eslint src/**/*.{js,jsx,ts,tsx,json}",
-        "lint:fix": "eslint --fix src/**/*.{js,jsx,ts,tsx,json}",
-        "format": "prettier --write src/**/*.{js,jsx,ts,tsx,css,md,json,scss} --config ./.prettierrc"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
+        "react-hook-form": "^7.45.4",
+        "react-router-dom": "^6.14.2"
     },
     "devDependencies": {
-        "@tailwindcss/forms": "^0.5.3",
-        "@vitest/coverage-c8": "^0.30.0",
-        "autoprefixer": "^10.4.14",
-        "eslint": "^8.37.0",
-        "eslint-config-airbnb": "^19.0.4",
-        "eslint-config-prettier": "^8.8.0",
-        "eslint-plugin-import": "^2.27.5",
-        "eslint-plugin-json": "^3.1.0",
-        "eslint-plugin-jsx-a11y": "^6.7.1",
-        "eslint-plugin-prettier": "^4.2.1",
-        "eslint-plugin-react": "^7.32.2",
-        "eslint-plugin-react-hooks": "^4.6.0",
-        "eslint-plugin-tailwindcss": "^3.11.0",
-        "eslint-plugin-testing-library": "^5.10.2",
-        "eslint-plugin-vitest": "^0.1.4",
-        "gh-pages": "^5.0.0",
-        "jsdom": "^21.1.1",
-        "postcss": "^8.4.21",
-        "prettier": "2.8.7",
-        "prettier-plugin-tailwindcss": "^0.2.7",
-        "tailwindcss": "^3.3.1"
+        "@testing-library/jest-dom": "^5.17.0",
+		"@testing-library/react": "^14.0.0",
+		"@testing-library/user-event": "^14.4.3",
+		"@types/react": "^18.2.20",
+		"@types/react-dom": "^18.2.7",
+		"@typescript-eslint/eslint-plugin": "^6.0.0",
+		"@typescript-eslint/parser": "^6.0.0",
+		"@vitejs/plugin-react": "^4.0.3",
+		"@vitejs/plugin-react-swc": "^3.3.2",
+		"autoprefixer": "^10.4.14",
+		"eslint": "^8.45.0",
+		"eslint-plugin-react-hooks": "^4.6.0",
+		"eslint-plugin-react-refresh": "^0.4.3",
+		"identity-obj-proxy": "^3.0.0",
+		"jest": "^29.6.2",
+		"jest-environment-jsdom": "^29.6.2",
+		"msw": "^1.2.3",
+		"postcss": "^8.4.27",
+		"tailwindcss": "^3.3.3",
+		"ts-jest": "^29.1.1",
+		"ts-node": "^10.9.1",
+		"typescript": "^5.0.2",
+		"vite": "^4.4.5",
+        "@tailwindcss/forms": "^0.5.3",     
+        "gh-pages": "^5.0.0"   
     }
 }
