@@ -3,10 +3,10 @@
 module.exports = {
     content: {
         relative: true,
-        files: ['./index.html', '././src/**/*.{js,jsx}', '././src/*.{js,jsx}'],
+        files: ['./index.html','./src/**/*.{js,ts,jsx,tsx}'],
     },
     theme: {
-        extend: {
+          extend: {
             backgroundImage: {
                 dragon: "url('/src/assets/dragon.svg')",
                 paper: "url('/src/assets/paper.svg')",
